+++ internal/pkg/consensus/block_validation.go
@@ -61,7 +61,7 @@ func NewDefaultBlockValidator(c clock.ChainEpochClock) *DefaultBlockValidator {
 // the chain clock.
 func (dv *DefaultBlockValidator) NotFutureBlock(b *block.Block) error {
 	currentEpoch := dv.EpochAtTime(dv.Now())
-	if types.NewBlockHeight(uint64(b.Height)).GreaterThan(currentEpoch) {
+	if types.NewBlockHeight(b.Height).GreaterThan(currentEpoch) {
 		return fmt.Errorf("block %s with timestamp %d generate in future epoch %d", b.Cid().String(), b.Timestamp, b.Height)
 	}
 	return nil
@@ -70,8 +70,8 @@ func (dv *DefaultBlockValidator) NotFutureBlock(b *block.Block) error {
 // TimeMatchesEpoch errors if the epoch and time don't match according to the
 // chain clock.
 func (dv *DefaultBlockValidator) TimeMatchesEpoch(b *block.Block) error {
-	earliestExpected, latestExpected := dv.EpochRangeAtTimestamp(uint64(b.Timestamp))
-	blockEpoch := types.NewBlockHeight(uint64(b.Height))
+	earliestExpected, latestExpected := dv.EpochRangeAtTimestamp(b.Timestamp)
+	blockEpoch := types.NewBlockHeight(b.Height)
 	if !earliestExpected.LessEqual(blockEpoch) || !latestExpected.GreaterEqual(blockEpoch) {
 		return fmt.Errorf(
 			"block %s with timestamp %d generated in wrong epoch %d, expected epoch in range [%d, %d]",
@@ -93,7 +93,7 @@ func (dv *DefaultBlockValidator) ValidateSemantic(ctx context.Context, child *bl
 		return err
 	}
 
-	if uint64(child.Height) <= ph {
+	if child.Height <= ph {
 		return fmt.Errorf("block %s has invalid height %d", child.Cid().String(), child.Height)
 	}
